pyproject.toml
[project]
name = "browse-use-on-cursor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use",
    "azure-ai-projects",
    "azure-identity",
    "python-dotenv",
    "fastapi",
    "uvicorn",
]
